Extract node matchers into separate module
